+++ brownie/_config.py
@@ -4,6 +4,7 @@ import json
 import os
 import re
 import shutil
+import sys
 import warnings
 from collections import defaultdict
 from pathlib import Path
@@ -15,12 +16,20 @@ from hypothesis.database import DirectoryBasedExampleDatabase
 
 from brownie._singleton import _Singleton
 
+__version__ = "1.9.3"
+
 BROWNIE_FOLDER = Path(__file__).parent
 DATA_FOLDER = Path.home().joinpath(".brownie")
 
 DATA_SUBFOLDERS = ("accounts", "ethpm", "packages")
 
 EVM_EQUIVALENTS = {"atlantis": "byzantium", "agharta": "petersburg"}
+
+python_version = (
+    f"{sys.version_info.major}.{sys.version_info.minor}"
+    f".{sys.version_info.micro} {sys.version_info.releaselevel}"
+)
+REQUEST_HEADERS = {"User-Agent": f"Brownie/{__version__} (Python/{python_version})"}
 
 
 class ConfigContainer: